refactor(generator): reorganize module structure
