+++ generator/src/main/kotlin/com/atls/generator/kmp/KmpModuleGenerator.kt
@@ -1,6 +1,6 @@
 package com.atls.generator.kmp
 
-import com.atls.generator.utils.ModuleGenerator
+import com.atls.generator.core.ModuleGenerator
 import org.gradle.api.Project
 
 class KmpModuleGenerator(
